fix: Move bot code
