+++ ferris_elf/bot.py
