+++ ferris_elf/constants.py
