+++ ops/systemd/ferris-elf-fetch.service
@@ -5,4 +5,4 @@ Description="Ferris Elf input fetcher"
 Type=oneshot
 EnvironmentFile=/srv/ferris-elf/config/env_file
 WorkingDirectory=/srv/ferris-elf/ferris-elf/
-ExecStart=/srv/ferris-elf/.local/bin/poetry run python3 fetch.py -d
+ExecStart=/srv/ferris-elf/.local/bin/poetry run python3 ferris_elf/fetch.py -d
